# openapi.yaml
openapi: 3.0.3
info:
  title: Rani Beauty Clinic API
  version: 1.0.0
  description: |
    Baseline API contract for production route surfaces.
    This spec is CI-validated and intended to evolve with the codebase.
  license:
    name: Proprietary
    url: https://rani-beauty-clinic.vercel.app
servers:
  - url: https://rani-beauty-clinic.vercel.app
    description: Production
...
security:
  - SessionCookie: []
paths:
...
  /api/dashboard/schedule/no-show-risk:
    get:
      tags: [Dashboard]
      operationId: getNoShowRisk
      summary: Get no-show risk scoring for upcoming appointments
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Optional single-day date filter.
      responses:
        "200":
          description: Ranked no-show risk list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoShowRiskItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/patient/auth/verify:
    post:
      tags: [Patient]
      operationId: postPatientVerifyMagicLink
      summary: Verify patient magic-link token and establish session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientVerifyRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Public mutation routes ───────────────────────────────────────
...
  /api/dashboard/communications/campaigns:
    get:
      operationId: listCampaigns
      summary: List all campaigns
      tags: [Dashboard]
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createCampaign
      summary: Create a campaign
      tags: [Dashboard]
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignCreateRequest"
      responses:
        "200":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Webhook routes ───────────────────────────────────────────────
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: rani-session
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    NoShowRiskItem:
      type: object
      properties:
        appointmentId:
          type: string
        clientName:
          type: string
        service:
          type: string
        provider:
          type: string
        date:
          type: string
        time:
          type: string
        noShowScore:
          type: object
          properties:
            score:
              type: number
            risk:
              type: string
          required: [score]
      required: [appointmentId, clientName, noShowScore]
...
    PatientVerifyRequest:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
          format: email
      required: [token, email]

    # ── Wave 4 named schemas ──────────────────────────────────────────
...
    CampaignCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [promotional, educational, reactivation, event, seasonal, birthday, direct]
        channel:
          type: string
          enum: [sms, email, both]
        subject:
          type: string
        body:
          type: string
      required: [name, type]
    CampaignResponse:
      type: object
      properties:
        success:
          type: boolean
        campaign:
          type: object
    CampaignListResponse:
      type: object
      properties:
        campaigns:
          type: array
          items:
            type: object
        total:
          type: integer
        asOf:
          type: string
          format: date-time

    # ── Webhook schemas ──────────────────────────────────────────────
...